branch listing and branch details for the CLI
