+++ crates/gitbutler-cli/Cargo.toml
@@ -15,6 +15,7 @@ gitbutler-oplog.workspace = true
 gitbutler-project.workspace = true
 gitbutler-reference.workspace = true
 gitbutler-branch-actions.workspace = true
+gitbutler-command-context.workspace = true
 gitbutler-branch.workspace = true
 gitbutler-diff.workspace = true
 gix = { workspace = true, features = ["max-performance-safe"] }
